+++ Changelog.md
@@ -63,6 +63,7 @@
 * Add parsing `DisplayLabelsProperties` in `Series`
 * Add parsing `FootnoteProperties`
 * Add parsing `Settings#default_tab_stop`
+* Add parsing `DocProperties`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit